requirements.txt
flask>=3.0.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
gunicorn>=21.2.0
